# Refactor ErrorHandler