@MadVikingGod 2021-12-01
Skipping changelog as this is a non-user facing change.